split murmur3 and hasher
