+++ src/engine/engine/murmur3/murmur3_128.js
@@ -1,24 +1,10 @@
 /* eslint-disable no-fallthrough */
-import utfx from 'utfx';
+import { UTF16ToUTF8, x86Fmix, x86Multiply, x86Rotl } from '.';
 
 const X64 = 'x64';
 const X86 = 'x86';
 const X64_ARCHS = ['arm64', 'ppc64', 'x64', 's390x', 'mipsel'];
-const stringSource = utfx.stringSource;
-const stringDestination = utfx.stringDestination;
-const encodeUTF16toUTF8 = utfx.encodeUTF16toUTF8;
 const isX64 = getArchType() === X64;
-
-function UTF16ToUTF8(key) {
-  let sd;
-
-  encodeUTF16toUTF8(
-    stringSource(key),
-    sd = stringDestination()
-  );
-
-  return sd();
-}
 
 /*!
  * +----------------------------------------------------------------------------------+
@@ -31,40 +17,6 @@ function UTF16ToUTF8(key) {
 
 // PRIVATE FUNCTIONS
 // -----------------
-function _x86Multiply(m, n) {
-  //
-  // Given two 32bit ints, returns the two multiplied together as a
-  // 32bit int.
-  //
-
-  return ((m & 0xffff) * n) + ((((m >>> 16) * n) & 0xffff) << 16);
-}
-
-
-function _x86Rotl(m, n) {
-  //
-  // Given a 32bit int and an int representing a number of bit positions,
-  // returns the 32bit int rotated left by that number of positions.
-  //
-
-  return (m << n) | (m >>> (32 - n));
-}
-
-function _x86Fmix(h) {
-  //
-  // Given a block, returns murmurHash3's final x86 mix of that block.
-  //
-
-  h ^= h >>> 16;
-  h = _x86Multiply(h, 0x85ebca6b);
-  h ^= h >>> 13;
-  h = _x86Multiply(h, 0xc2b2ae35);
-  h ^= h >>> 16;
-
-  return h;
-}
-
-
 function _x64Add(m, n) {
   //
   // Given two 64bit ints (as an array of two 32bit ints) returns the two
@@ -197,59 +149,6 @@ function _x64Fmix(h) {
 
 // PUBLIC FUNCTIONS
 // ----------------
-function hash32(key, seed) {
-  //
-  // Given a string and an optional seed as an int, returns a 32 bit hash
-  // using the x86 flavor of MurmurHash3, as an unsigned int.
-  //
-
-  key = key || '';
-  seed = seed || 0;
-
-  var remainder = key.length % 4;
-  var bytes = key.length - remainder;
-
-  var h1 = seed;
-
-  var k1 = 0;
-
-  var c1 = 0xcc9e2d51;
-  var c2 = 0x1b873593;
-
-  for (var i = 0; i < bytes; i = i + 4) {
-    k1 = ((key.charCodeAt(i) & 0xff)) | ((key.charCodeAt(i + 1) & 0xff) << 8) | ((key.charCodeAt(i + 2) & 0xff) << 16) | ((key.charCodeAt(i + 3) & 0xff) << 24);
-
-    k1 = _x86Multiply(k1, c1);
-    k1 = _x86Rotl(k1, 15);
-    k1 = _x86Multiply(k1, c2);
-
-    h1 ^= k1;
-    h1 = _x86Rotl(h1, 13);
-    h1 = _x86Multiply(h1, 5) + 0xe6546b64;
-  }
-
-  k1 = 0;
-
-  switch (remainder) {
-    case 3:
-      k1 ^= (key.charCodeAt(i + 2) & 0xff) << 16;
-    case 2:
-      k1 ^= (key.charCodeAt(i + 1) & 0xff) << 8;
-
-    case 1:
-      k1 ^= (key.charCodeAt(i) & 0xff);
-      k1 = _x86Multiply(k1, c1);
-      k1 = _x86Rotl(k1, 15);
-      k1 = _x86Multiply(k1, c2);
-      h1 ^= k1;
-  }
-
-  h1 ^= key.length;
-  h1 = _x86Fmix(h1);
-
-  return h1 >>> 0;
-}
-
 function hash128x86(key, seed) {
   //
   // Given a string and an optional seed as an int, returns a 128 bit
@@ -283,41 +182,41 @@ function hash128x86(key, seed) {
     k3 = ((key.charCodeAt(i + 8) & 0xff)) | ((key.charCodeAt(i + 9) & 0xff) << 8) | ((key.charCodeAt(i + 10) & 0xff) << 16) | ((key.charCodeAt(i + 11) & 0xff) << 24);
     k4 = ((key.charCodeAt(i + 12) & 0xff)) | ((key.charCodeAt(i + 13) & 0xff) << 8) | ((key.charCodeAt(i + 14) & 0xff) << 16) | ((key.charCodeAt(i + 15) & 0xff) << 24);
 
-    k1 = _x86Multiply(k1, c1);
-    k1 = _x86Rotl(k1, 15);
-    k1 = _x86Multiply(k1, c2);
+    k1 = x86Multiply(k1, c1);
+    k1 = x86Rotl(k1, 15);
+    k1 = x86Multiply(k1, c2);
     h1 ^= k1;
 
-    h1 = _x86Rotl(h1, 19);
+    h1 = x86Rotl(h1, 19);
     h1 += h2;
-    h1 = _x86Multiply(h1, 5) + 0x561ccd1b;
+    h1 = x86Multiply(h1, 5) + 0x561ccd1b;
 
-    k2 = _x86Multiply(k2, c2);
-    k2 = _x86Rotl(k2, 16);
-    k2 = _x86Multiply(k2, c3);
+    k2 = x86Multiply(k2, c2);
+    k2 = x86Rotl(k2, 16);
+    k2 = x86Multiply(k2, c3);
     h2 ^= k2;
 
-    h2 = _x86Rotl(h2, 17);
+    h2 = x86Rotl(h2, 17);
     h2 += h3;
-    h2 = _x86Multiply(h2, 5) + 0x0bcaa747;
+    h2 = x86Multiply(h2, 5) + 0x0bcaa747;
 
-    k3 = _x86Multiply(k3, c3);
-    k3 = _x86Rotl(k3, 17);
-    k3 = _x86Multiply(k3, c4);
+    k3 = x86Multiply(k3, c3);
+    k3 = x86Rotl(k3, 17);
+    k3 = x86Multiply(k3, c4);
     h3 ^= k3;
 
-    h3 = _x86Rotl(h3, 15);
+    h3 = x86Rotl(h3, 15);
     h3 += h4;
-    h3 = _x86Multiply(h3, 5) + 0x96cd1c35;
+    h3 = x86Multiply(h3, 5) + 0x96cd1c35;
 
-    k4 = _x86Multiply(k4, c4);
-    k4 = _x86Rotl(k4, 18);
-    k4 = _x86Multiply(k4, c1);
+    k4 = x86Multiply(k4, c4);
+    k4 = x86Rotl(k4, 18);
+    k4 = x86Multiply(k4, c1);
     h4 ^= k4;
 
-    h4 = _x86Rotl(h4, 13);
+    h4 = x86Rotl(h4, 13);
     h4 += h1;
-    h4 = _x86Multiply(h4, 5) + 0x32ac3b17;
+    h4 = x86Multiply(h4, 5) + 0x32ac3b17;
   }
 
   k1 = 0;
@@ -334,9 +233,9 @@ function hash128x86(key, seed) {
 
     case 13:
       k4 ^= key.charCodeAt(i + 12);
-      k4 = _x86Multiply(k4, c4);
-      k4 = _x86Rotl(k4, 18);
-      k4 = _x86Multiply(k4, c1);
+      k4 = x86Multiply(k4, c4);
+      k4 = x86Rotl(k4, 18);
+      k4 = x86Multiply(k4, c1);
       h4 ^= k4;
 
     case 12:
@@ -350,9 +249,9 @@ function hash128x86(key, seed) {
 
     case 9:
       k3 ^= key.charCodeAt(i + 8);
-      k3 = _x86Multiply(k3, c3);
-      k3 = _x86Rotl(k3, 17);
-      k3 = _x86Multiply(k3, c4);
+      k3 = x86Multiply(k3, c3);
+      k3 = x86Rotl(k3, 17);
+      k3 = x86Multiply(k3, c4);
       h3 ^= k3;
 
     case 8:
@@ -366,9 +265,9 @@ function hash128x86(key, seed) {
 
     case 5:
       k2 ^= key.charCodeAt(i + 4);
-      k2 = _x86Multiply(k2, c2);
-      k2 = _x86Rotl(k2, 16);
-      k2 = _x86Multiply(k2, c3);
+      k2 = x86Multiply(k2, c2);
+      k2 = x86Rotl(k2, 16);
+      k2 = x86Multiply(k2, c3);
       h2 ^= k2;
 
     case 4:
@@ -382,9 +281,9 @@ function hash128x86(key, seed) {
 
     case 1:
       k1 ^= key.charCodeAt(i);
-      k1 = _x86Multiply(k1, c1);
-      k1 = _x86Rotl(k1, 15);
-      k1 = _x86Multiply(k1, c2);
+      k1 = x86Multiply(k1, c1);
+      k1 = x86Rotl(k1, 15);
+      k1 = x86Multiply(k1, c2);
       h1 ^= k1;
   }
 
@@ -400,10 +299,10 @@ function hash128x86(key, seed) {
   h3 += h1;
   h4 += h1;
 
-  h1 = _x86Fmix(h1);
-  h2 = _x86Fmix(h2);
-  h3 = _x86Fmix(h3);
-  h4 = _x86Fmix(h4);
+  h1 = x86Fmix(h1);
+  h2 = x86Fmix(h2);
+  h3 = x86Fmix(h3);
+  h4 = x86Fmix(h4);
 
   h1 += h2;
   h1 += h3;
@@ -534,14 +433,6 @@ function hash128x64(key, seed) {
   return ('00000000' + (h1[0] >>> 0).toString(16)).slice(-8) + ('00000000' + (h1[1] >>> 0).toString(16)).slice(-8) + ('00000000' + (h2[0] >>> 0).toString(16)).slice(-8) + ('00000000' + (h2[1] >>> 0).toString(16)).slice(-8);
 }
 
-function hash(str /*: string */, seed /*: number */) /*: number */ {
-  return hash32(UTF16ToUTF8(str), seed >>> 0);
-}
-
-function bucket(str /*: string */, seed /*: number */) /*: number */ {
-  return Math.abs(hash(str, seed) % 100) + 1;
-}
-
 function getArchType() {
   // Values listed from https://nodejs.org/api/process.html#process_process_arch
   // @TODO Review when supporting alternatives of Node
@@ -549,12 +440,6 @@ function getArchType() {
   return (X64_ARCHS.indexOf(arch) > -1) ? X64 : X86;
 }
 
-function hash128(str /*: string */, seed /*: number */) /*: string */ {
+export function hash128(str /*: string */, seed /*: number */) /*: string */ {
   return isX64 ? hash128x64(UTF16ToUTF8(str), seed >>> 0) : hash128x86(UTF16ToUTF8(str), seed >>> 0);
 }
-
-export default {
-  hash,
-  hash128,
-  bucket
-};
